Task 2 first draft

+++ models/student.go
@@ -1,7 +1,7 @@
 package models
 
 type Student struct {
-	RollNo       int64
+	RollNo       int
 	Name         string `binding:"required"`
 	GuardianName string `binding:"required"`
 	Address      string `binding:"required"`
